Ultimos detalles de la practica de los Fireworks
Se lanzan por tiempo, en la practica de fuerzas se han sustituido cpps obsoletos

+++ Física/Fuerzas/Practica1/Firework.h
@@ -9,7 +9,7 @@ private:
 	unsigned type_;
 
 	//tiempo en explotar
-	double age_;
+	float age_;
 
 public:
 	Firework() {};
